// src/shared/generated/EmbeddedRuntimeWindowProjectionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EmbeddedRuntimeTabPhaseProjectionRecord } from "./EmbeddedRuntimeTabPhaseProjectionRecord";
import type { EmbeddedRuntimeWorkspaceTabProjectionRecord } from "./EmbeddedRuntimeWorkspaceTabProjectionRecord";

export type EmbeddedRuntimeWindowProjectionRecord = { windowId: string, windowGeneration: number, topologyRevision: number, tabIds: Array<string>, tabPhases: Array<EmbeddedRuntimeTabPhaseProjectionRecord>, hiddenTabIds: Array<string>, workspaceTabs?: Array<EmbeddedRuntimeWorkspaceTabProjectionRecord>, activeTabId?: string, };