// src/shared/generated/EmbeddedRuntimeWorkspaceTabProjectionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StateWorkspaceSlotRecord } from "./StateWorkspaceSlotRecord";

export type EmbeddedRuntimeWorkspaceTabProjectionRecord = { tabId: string, workspaceSlots: Array<StateWorkspaceSlotRecord>, };